requirements.txt
# Core Dependencies
pandas>=1.5.0
kafka-python>=2.0.2
pyspark==3.3.0
pymongo>=4.3.3
streamlit>=1.28.0

# Web Framework
fastapi>=0.104.0
uvicorn>=0.24.0

# Development & Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.1
pytest-benchmark>=4.0.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.5.0

# Security
safety>=2.3.0
bandit>=1.7.5

# Monitoring & Logging
python-dotenv>=1.0.0
structlog>=23.1.0

# Performance
numpy>=1.24.0
psutil>=5.9.0

# HTTP Client (for testing)
requests>=2.31.0

# Optional: Data visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0